requirements_V3.txt
bento-mdf==0.10.0
boto3==1.37.32
botocore==1.37.32
graphql-core==3.2.6
greenlet==3.1.1
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
importlib_metadata==8.6.1
jmespath==1.0.1
neo4j==5.28.1
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.3
pip-autoremove==0.10.0
prefect-aws==0.5.9
prefect-docker==0.6.2
prefect-github==0.3.1
prefect-shell==0.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
s3transfer==0.11.4
sgqlc==16.4
six==1.17.0
tabulate==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
zipp==3.21.0
